Update kube package handlers to be multi-cluster aware.
Moves NewClusterConfig (and tests) into the shared kube package so that
it can be used by other packages. Updates the AsUser helper (which
returns clientsets) to use NewClusterConfig, requiring a call-signature
change.

I'll followup with a related small fix to ensure we don't swallow errors
in AsUser.
